+++ grammar/grammar.py
@@ -169,8 +169,7 @@ class LangDef(Grammar):
 
     return_statement = Sequence(
         k_return,
-        THIS,
-        Optional(Sequence(',', THIS)))
+        List(THIS, mi=1, ma=3, opt=False))
 
     for_statement = Sequence(
         k_for,