+++ inc/langdef/langdef.h
@@ -5,7 +5,7 @@
  * should be used with the libcleri module.
  *
  * Source class: LangDef
- * Created at: 2022-06-28 11:31:22
+ * Created at: 2022-08-03 14:34:20
  */
 #ifndef CLERI_EXPORT_LANGDEF_H_
 #define CLERI_EXPORT_LANGDEF_H_