+++ inc/ti/thing.h
@@ -78,7 +78,11 @@ int ti_thing_get_by_raw_e(
         ti_raw_t * r,
         ex_t * e);
 int ti_thing_gen_id(ti_thing_t * thing);
-int ti_thing__to_client_pk(ti_thing_t * thing, ti_vp_t * vp, int deep);
+int ti_thing__to_client_pk(
+        ti_thing_t * thing,
+        ti_vp_t * vp,
+        int deep,
+        int flags);
 int ti_thing_o_to_pk(ti_thing_t * thing, msgpack_packer * pk);
 int ti_thing_t_to_pk(ti_thing_t * thing, msgpack_packer * pk);
 ti_val_t * ti_thing_val_by_strn(ti_thing_t * thing, const char * str, size_t n);
@@ -152,6 +156,16 @@ static inline _Bool ti_thing_is_instance(ti_thing_t * thing)
 }
 
 static inline int ti_thing_id_to_client_pk(ti_thing_t * thing, msgpack_packer * pk)
+{
+    assert (thing->id);
+    return -(
+            msgpack_pack_map(pk,1) ||
+            mp_pack_strn(pk, TI_KIND_S_THING, 1) ||
+            msgpack_pack_uint64(pk, thing->id)
+    );
+}
+
+static inline int ti_thing_empty_to_client_pk(ti_thing_t * thing, msgpack_packer * pk)
 {
     return -(msgpack_pack_map(pk, !!thing->id) ||
         (thing->id && (