+++ inc/ti/thing.inline.h
@@ -147,11 +147,14 @@ static inline ti_val_t * ti_thing_val_weak_by_name(
 static inline int ti_thing_to_client_pk(
         ti_thing_t * thing,
         ti_vp_t * vp,
-        int deep)
+        int deep,
+        int flags)
 {
-    return !deep || (thing->flags & TI_VFLAG_LOCK)
-            ? ti_thing_id_to_client_pk(thing, &vp->pk)
-            : ti_thing__to_client_pk(thing, vp, deep);
+    return (!deep || (thing->flags & TI_VFLAG_LOCK))
+            ? (!thing->id || (flags & TI_FLAGS_NO_IDS))
+            ? ti_thing_empty_to_client_pk(thing, &vp->pk)
+            : ti_thing_id_to_client_pk(thing, &vp->pk)
+            : ti_thing__to_client_pk(thing, vp, deep, flags);
 }
 
 static inline int ti_thing_to_store_pk(ti_thing_t * thing, msgpack_packer * pk)