+++ itest/test_statements.py
@@ -237,6 +237,32 @@ class TestStatements(TestBase):
         """)
         self.assertEqual(res, {'c': 2})
 
+    async def test_return_flags(self, client):
+        res = await client.query("""//ti
+            set_type('A', {
+                nested: 'thing'
+            });
+            .a = A();
+            return .a, 2, NO_IDS;
+        """)
+        self.assertEqual(res, {'nested': {}})
+
+        res = await client.query("""//ti
+            return .a, 2, 1;
+        """)
+        self.assertIn('#', res)
+        self.assertIn('#', res['nested'])
+
+        res = await client.query("""//ti
+            str(|| return 'value',0,NO_IDS);
+        """)
+        self.assertEqual(res, "|| return 'value', 0, NO_IDS")
+
+        res = await client.query("""//ti
+            || return 'value',0,NO_IDS;
+        """)
+        self.assertEqual(res, "|| return 'value',0,NO_IDS")
+
     async def test_format(self, client):
         res = await client.query("""//ti
             x = || {